Fix up zipfile logic
