+++ app/services/data_exports/create_service.rb
@@ -50,6 +50,7 @@ module DataExports
       project_ids.each do |project_id|
         authorize! Project.find(project_id), to: :export_sample_data?
       end
+      projects
     end
   end
 end
